pyproject.toml
[project]
name = "webapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.22.0",
    "aiosqlite>=0.21.0",
    "fastapi>=0.119.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
